Add files via upload
Here is the python code that converts symbol frame like "1101010010101xxxx" to a specific voltage levels. The waveform then can be extracted to a csv file. For example, the pattern "1101000111010101010111100001011" is converted w.r.t time analog signal in which symbol 1 is represented by 0.65V level and symbol 0 is represented by 0.45V level. The symbol duration is 1ms and both the rise and fall times are 100us. The X-axi sis the time scale and Y-axis is the symbol levels.